finish styles for login and signup pages
