--- src/components/Form/Form.css
@@ -1,43 +1,0 @@
-.app__form {
-  max-width: 600px;
-}
-
-.app__form form {
-  background-color: rgba(61, 61, 61, 0.827);
-  padding: 30px 60px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 3rem;
-  border-radius: 10px;
-}
-
-.app__form-h1 {
-  margin: 1rem 0;
-  font-family: var(--font-alt);
-  color: var(--color-golden);
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  font-size: 30px;
-  text-align: center;
-}
-
-.app__form-button {
-  width: 300px;
-  height: 50px;
-  padding: 10px;
-  border: none;
-  background-color: #dcca8762;
-  color: white;
-  border-radius: 5px;
-  font-weight: bold;
-  font-size: 18px;
-  cursor: pointer;
-  margin-top: 20px;
-  margin-bottom: 30px;
-  transition: 0.3s;
-}
-
-.app__form-button:hover {
-  background-color: #dcca8798;
-}
