+++ src/pages/SignUp/SignUp.css
@@ -1,11 +1,36 @@
 .app__signup {
   width: 100%;
   height: fit-content;
+  min-height: 100vh;
   display: flex;
   background: rgba(26, 26, 26, 0.944);
   padding: 1rem 10rem;
   align-items: center;
   flex-direction: column;
+}
+
+.app__signup-form {
+  max-width: 600px;
+}
+
+.app__signup-form form {
+  background-color: rgba(61, 61, 61, 0.827);
+  padding: 30px 60px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 3rem;
+  border-radius: 10px;
+}
+
+.app__signup-form-h1 {
+  margin: 1rem 0;
+  font-family: var(--font-alt);
+  color: var(--color-golden);
+  letter-spacing: 0.04em;
+  text-transform: uppercase;
+  font-size: 30px;
+  text-align: center;
 }
 
 .app__signup-redirect {
@@ -16,4 +41,24 @@
 
 .app__signup-redirect-span {
   text-decoration: underline;
+}
+
+.app__form-button {
+  width: 300px;
+  height: 50px;
+  padding: 10px;
+  border: none;
+  background-color: #dcca8762;
+  color: white;
+  border-radius: 5px;
+  font-weight: bold;
+  font-size: 18px;
+  cursor: pointer;
+  margin-top: 20px;
+  margin-bottom: 30px;
+  transition: 0.3s;
+}
+
+.app__form-button:hover {
+  background-color: #dcca8798;
 }
